Fix some minor bugs
Found/fix a couple of bugs related to the program:
- When running this program on a Linux/Unix based machine all file
passed in by the user would failed because of the fact that the path
separator was hard-coded in for that file location. To fix the problem
we now use python's built-in way of generating file paths.
- If the user wish to only generate a blank madlib export it would
always fail. The problem was caused by the fact that the equality
was checking for the int value of the user's choice instead of the
string value.